use fake app storage in explorer_service_test.dart
